// FrontEnd/src/Admin/models/ImportModel.ts
export interface ImportModel {
    id: string;
    partner_name: string;
    create_by_name: string;
    note: string;
    total: string;
    created_at: string;
    updated_at: string;
}
export interface ImportDetailModel {
    id: string;
    product_name: string;
    product_img: string;
    quantity: string;
    price: string;
    total: string;
}